+++ lib/index.ts
@@ -7,7 +7,8 @@ export {
   uuid_to_string, generate_uuid, string_to_uuid, LinkError, ProtocolError, LinkOptions,
   DeliveryAnnotations, MessageAnnotations, ReceiverEvents, SenderEvents, ConnectionEvents,
   SessionEvents, ContainerOptions as ContainerOptionsBase, TerminusOptions, Types, Sasl,
-  EndpointOptions, MessageUtil, TypeError, SimpleError, Source, ConnectionError, Typed, WebSocketImpl, WebSocketInstance
+  EndpointOptions, MessageUtil, TypeError, SimpleError, Source, ConnectionError, Typed,
+  WebSocketImpl, WebSocketInstance, TargetTerminusOptions
 } from "rhea";
 
 export { EventContext, OnAmqpEvent } from "./eventContext";
@@ -18,6 +19,7 @@ export {
 export { Session } from "./session";
 export { Receiver, ReceiverOptions } from "./receiver";
 export { Sender, SenderOptions } from "./sender";
+export { LinkCloseOptions } from "./link";
 export {
   Func, AmqpResponseStatusCode, isAmqpError, ConnectionStringParseOptions, delay, messageHeader,
   messageProperties, parseConnectionString, ParsedOutput